Feat: content fixed in mobile menu
